minio-autoscript/requirements.txt
minio~=7.2.0
paho-mqtt~=1.6.1
Pillow~=10.1.0
